+++ src/frgp/utn/edu/ar/negocio/ITurnoNegocio.java
@@ -7,7 +7,7 @@ public interface ITurnoNegocio {
 
 	    boolean add(Turno turno);
 
-	    Turno readOne(Long id);
+	    Turno readOne(int id);
 
 	    List<Turno> readAll();
